maint: Replace float by double
* src/ctxmgmt/SystemInfo.cpp: Likewise.
* src/ctxmgmt/SystemInfo.h: Likewise.
* src/formatter/AdapterFormatterPlayer.cpp: Likewise.
* src/formatter/AdapterNCLPlayer.cpp: Likewise.
* src/formatter/NclFormatterLayout.cpp: Likewise.
* src/formatter/NclFormatterLayout.h: Likewise.
* src/formatter/NclFormatterRegion.cpp: Likewise.
* src/formatter/NclFormatterRegion.h: Likewise.
* src/ic/CurlInteractiveChannel.cpp: Likewise.
* src/ic/CurlInteractiveChannel.h: Likewise.
* src/mb/DisplayManager.cpp: Likewise.
* src/mb/DisplayManager.h: Likewise.
* src/mb/IContinuousMediaProvider.h: Likewise.
* src/mb/SDL2ffmpeg.cpp: Likewise.
* src/mb/SDL2ffmpeg.h: Likewise.
* src/mb/SDLAudioProvider.cpp: Likewise.
* src/mb/SDLAudioProvider.h: Likewise.
* src/mb/SDLDeviceScreen.cpp: Likewise.
* src/mb/SDLDeviceScreen.h: Likewise.
* src/mb/SDLVideoProvider.cpp: Likewise.
* src/mb/SDLVideoProvider.h: Likewise.
* src/mb/SDLWindow.cpp: Likewise.
* src/mb/SDLWindow.h: Likewise.
* src/ncl/Comparator.cpp: Likewise.
* src/ncl/Comparator.h: Likewise.
* src/ncl/LayoutRegion.cpp: Likewise.
* src/nclconv/NclLayoutConverter.cpp: Likewise.
* src/nclconv/NclLayoutConverter.h: Likewise.
* src/player/AVPlayer.cpp: Likewise.
* src/player/AVPlayer.h: Likewise.
* src/player/INCLPlayer.h: Likewise.
* src/player/SrtPlayer.cpp: Likewise.
* src/player/SrtPlayer.h: Likewise.
* src/util/functions.cpp: Likewise.
* src/util/functions.h: Likewise.

+++ src/ctxmgmt/SystemInfo.h
@@ -25,7 +25,7 @@ GINGA_CTXMGMT_BEGIN
 class SystemInfo
 {
 private:
-  float clockSpeed;
+  double clockSpeed;
   map<string, string> *sysTable;
 
 public:
@@ -42,13 +42,13 @@ public:
   string getSystemLanguage ();
   string getCaptionLanguage ();
   string getSubtitleLanguage ();
-  float getReturnBitRate ();
+  double getReturnBitRate ();
   void getScreenSize (GingaScreenID screenId, int *width, int *height);
   void getScreenGraphicSize (GingaScreenID screenId, int *width,
                              int *height);
   string getAudioType ();
-  float getCPUClock ();
-  float getMemorySize ();
+  double getCPUClock ();
+  double getMemorySize ();
   string getJavaConfiguration ();
   string getJavaProfile ();
   string getLuaVersion ();
